exp/amake/requirements.txt
AssetForge>=0.2.6
Pillow>=8.0.0
numpy
opencv-python>=4.5.0
trimesh
scipy
flatbuffers